business_material 테이블 amount 컬럼 타입 변경 INT -> DECIMAL(7,2)

+++ src/main/java/com/interior/adapter/inbound/business/webdto/CreateBusinessMaterial.java
@@ -1,5 +1,6 @@
 package com.interior.adapter.inbound.business.webdto;
 
+import java.math.BigDecimal;
 import lombok.ToString;
 
 public class CreateBusinessMaterial {
@@ -8,7 +9,7 @@ public class CreateBusinessMaterial {
         String materialName,
         String materialUsageCategory,
         String materialCategory,
-        int materialAmount,
+        BigDecimal materialAmount,
         String materialAmountUnit,
         String materialMemo,
         String materialCostPerUnit,
